One way of cross-compiling local binary packages
If $METEOR_BINARY_DEP_WORKAROUND is set, then when bundling for a
non-host platform, if a package has no server unibuild for the target
architecture, use the host architecture and replace the npm modules with
a package.json and npm-shrinkwrap.json.  Also write out a top-level
setup.sh script (inside programs/server) which runs npm install in all
such directories.

To support this, we make sure to save the package.json and
npm-shrinkwrap.json files in various intermediate directories in case we
need them later.  (We put them inside node_modules because that is what
gets copied from source tree to isopack.)

Note: if you have a previously published package which has a unibuild
for host but not for server target and try to build, it will probably
crash.  You should publish-for-arch in that case, or wait for somebody
to publish a new version of the package with the patch to meteor-npm.js
from this file.